move InitializedObjectConstructor to src directory
