+++ src/Construction/InitializedObjectConstructor.php
@@ -16,24 +16,20 @@
  * limitations under the License.
  */
 
-namespace JMS\Serializer\Tests\Fixtures;
+namespace JMS\Serializer\Construction;
 
-use JMS\Serializer\Construction\ObjectConstructorInterface;
 use JMS\Serializer\DeserializationContext;
 use JMS\Serializer\Metadata\ClassMetadata;
 use JMS\Serializer\VisitorInterface;
 
 /**
- * Object constructor that allows deserialization into already constructed
- * objects passed through the deserialization context
+ * Object constructor that allows deserialization into already constructed objects passed through the deserialization context.
  */
 class InitializedObjectConstructor implements ObjectConstructorInterface
 {
     private $fallbackConstructor;
 
     /**
-     * Constructor.
-     *
      * @param ObjectConstructorInterface $fallbackConstructor Fallback object constructor
      */
     public function __construct(ObjectConstructorInterface $fallbackConstructor)
@@ -52,5 +48,4 @@ class InitializedObjectConstructor implements ObjectConstructorInterface
 
         return $this->fallbackConstructor->construct($visitor, $metadata, $data, $type, $context);
     }
-
 }
