+++ tests/Serializer/BaseSerializationTest.php
@@ -22,6 +22,7 @@ use Doctrine\Common\Annotations\AnnotationReader;
 use Doctrine\Common\Collections\ArrayCollection;
 use JMS\Serializer\Accessor\DefaultAccessorStrategy;
 use JMS\Serializer\Accessor\ExpressionAccessorStrategy;
+use JMS\Serializer\Construction\InitializedObjectConstructor;
 use JMS\Serializer\Construction\UnserializeObjectConstructor;
 use JMS\Serializer\Context;
 use JMS\Serializer\DeserializationContext;
@@ -69,7 +70,6 @@ use JMS\Serializer\Tests\Fixtures\GroupsObject;
 use JMS\Serializer\Tests\Fixtures\GroupsUser;
 use JMS\Serializer\Tests\Fixtures\IndexedCommentsBlogPost;
 use JMS\Serializer\Tests\Fixtures\InitializedBlogPostConstructor;
-use JMS\Serializer\Tests\Fixtures\InitializedObjectConstructor;
 use JMS\Serializer\Tests\Fixtures\InlineChild;
 use JMS\Serializer\Tests\Fixtures\InlineChildEmpty;
 use JMS\Serializer\Tests\Fixtures\InlineChildWithGroups;
